Documentation now shows what I will do next time I rebuild this script. For now the script is only 0.01 level, but it contains the basic skeleton I need to build what I wish to build. The thing I wish to build is specc'd in the html docs for the Line6PodXTFloor.html.
The control script itself is doing almost nothing when you press A,B,C,D, and the functions for F1-F5 are just some initial stabs aimed at learning the way this works.

The critical limitation I'm seeing is that I don't see a way to get notified in  the script about the change in track selection.  Perhaps when a column is selected, I can also detect a track change.